security

+++ src/Controller/EventController.php
@@ -49,8 +49,10 @@ final class EventController extends AbstractController
      * @throws TransportExceptionInterface
      */
     #[Route('/create', name: '_create')]
-    public function create(Request $request, EntityManagerInterface $em, ParameterBagInterface $parameterBag, FileUploader $fileUploader, #[CurrentUser] ?User $user, Security $security, GroupRepository $groupRepository, MailerInterface $mailer): Response
+    public function create(Request $request, EntityManagerInterface $em, ParameterBagInterface $parameterBag, FileUploader $fileUploader, #[CurrentUser] ?User $user, GroupRepository $groupRepository, MailerInterface $mailer): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $event = new Event();
         $place = new Place();
 
@@ -138,7 +140,10 @@ final class EventController extends AbstractController
     #[Route('/edit/{id}', name: '_edit', requirements: ['id' => '\d+'])]
     public function edit(Event $event, Request $request, EntityManagerInterface $em, ParameterBagInterface $parameterBag, FileUploader $fileUploader, Security $security): Response
     {
-        $this->checkStatusUser($event, $security);
+        if ($redirect = $this->checkStatusUser($event, $security))
+        {
+           return $redirect;
+        };
 
         $form = $this->createForm(EventType::class, $event);
         $placeForm = $this->createForm(PlaceType::class, new Place());
@@ -179,6 +184,8 @@ final class EventController extends AbstractController
         EntityManagerInterface $entityManager
     ): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $limit = $bag->get('event')['nb_max'];
         $offset = ($page - 1) * $limit;
 
@@ -274,14 +281,17 @@ final class EventController extends AbstractController
      * @throws Exception
      */
     #[Route('/detail/{id}', name: '_detail', requirements: ['id' => '\d+'])]
-    public function detail(SortieRepository $sortieRepository, int $id, ParameterBagInterface $bag, #[CurrentUser] ?User $userConnected): Response
+    public function detail(SortieRepository $sortieRepository, int $id, #[CurrentUser] ?User $userConnected): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $event = $sortieRepository->find($id);
         $userConnectedId = $userConnected->getId();
 
 
         if (!$event) {
-            throw $this->createNotFoundException('Cet événement n\'existe pas');
+            $this->addFlash('error','Cet évènement n\'existe pas');
+            return $this->redirectToRoute('event_list');
         }
 
         $participants = $sortieRepository->findParticipantsByEvent($event->getId());
@@ -298,7 +308,7 @@ final class EventController extends AbstractController
      * @throws Exception
      */
 
-    public function closeIfFullParticipants(StateRepository $stateRepository, SortieRepository $sortieRepository, int $id, ParameterBagInterface $bag, EntityManagerInterface $entityManager): void
+    public function closeIfFullParticipants(StateRepository $stateRepository, SortieRepository $sortieRepository, int $id, EntityManagerInterface $entityManager): void
     {
         $event = $sortieRepository->find($id);
         $participants = $sortieRepository->findParticipantsByEvent($event->getId());
@@ -314,7 +324,8 @@ final class EventController extends AbstractController
         }
     }
 
-    public function closeIfOutDate(StateRepository $stateRepository, SortieRepository $sortieRepository, int $id, ParameterBagInterface $bag, EntityManagerInterface $entityManager): void
+    #todo non utilisé à supprimer?
+    /*public function closeIfOutDate(StateRepository $stateRepository, SortieRepository $sortieRepository, int $id, EntityManagerInterface $entityManager): void
     {
         $event = $sortieRepository->find($id);
         $today = new \DateTime();
@@ -326,7 +337,7 @@ final class EventController extends AbstractController
             $entityManager->persist($event);
             $entityManager->flush();
         }
-    }
+    }*/
 
     /**
      * @throws Exception
@@ -334,13 +345,15 @@ final class EventController extends AbstractController
     #[Route('/join/{id}', name: '_join', requirements: ['id' => '\d+'])]
     public function join(StateRepository $stateRepository, SortieRepository $sortieRepository, #[CurrentUser] ?User $userConnected, int $id, ParameterBagInterface $bag, EntityManagerInterface $entityManager, MailerInterface $mailer, LoggerInterface $logger, MessageBusInterface $bus): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
 
         $event = $sortieRepository->find($id);
         $participants = $sortieRepository->findParticipantsByEvent($event->getId());
         $user= $userConnected->getId();
 
         if ($event->getState()->getId() !== 2) {
-            throw $this->createAccessDeniedException("Tu ne peux pas t'inscrire à cet événement");
+            $this->addFlash('error',"Tu ne peux pas t'inscrire à cet évènement");
+            return $this->redirectToRoute('event_list');
         }
 
         $nbParticipants = count($participants);
@@ -411,6 +424,8 @@ final class EventController extends AbstractController
         MailerInterface       $mailer
     ): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $event = $sortieRepository->find($id);
         if (!$event) {
             $this->addFlash('danger', "Événement introuvable.");
@@ -462,7 +477,10 @@ final class EventController extends AbstractController
     #[Route ('/cancel/{id}', name: '_cancel', requirements: ['id' => '\d+'])]
     public function cancel(Event $event, EntityManagerInterface $em, Security $security, StateRepository $stateRepository, Request $request): Response
     {
-        $this->checkStatusUser($event, $security);
+        if ($redirect = $this->checkStatusUser($event, $security))
+        {
+            return $redirect;
+        };
 
         $form = $this->createForm(CancellationReasonType::class, $event);
         $form->handleRequest($request);
@@ -488,7 +506,10 @@ final class EventController extends AbstractController
     #[Route ('/reactivate/{id}', name: '_reactivate', requirements: ['id' => '\d+'])]
     public function reactivate(Event $event, EntityManagerInterface $em, Security $security, StateRepository $stateRepository): Response
     {
-        $this->checkStatusUser($event, $security);
+        if ($redirect = $this->checkStatusUser($event, $security))
+        {
+            return $redirect;
+        };
 
         $reac = $stateRepository->findOneBy(['label' => 'Ouverte']);
         if (!$reac) {
@@ -504,8 +525,11 @@ final class EventController extends AbstractController
     #[Route('/delete/{id}', name: '_delete', requirements: ['id' => '\d+'])]
     public function delete(Event $event, Request $request, EntityManagerInterface $em, Security $security): Response
     {
+        if ($redirect = $this->checkStatusUser($event, $security))
+        {
+            return $redirect;
+        };
 
-        $this->checkStatusUser($event, $security);
         if ($this->isCsrfTokenValid('delete' . $event->getId(), $request->get('token'))) {
 
             $em->remove($event);
@@ -519,15 +543,12 @@ final class EventController extends AbstractController
         return $this->redirectToRoute('event_list');
     }
 
-    private function checkStatusUser(Event $event, Security $security): void
+    private function checkStatusUser(Event $event, Security $security): Response
     {
-
         if ($event->getOrganizer() !== $security->getUser() && !$security->isGranted('ROLE_ADMIN')) {
-            throw $this->createAccessDeniedException("Tu n'es pas l'organisateur de cet événement");
+            $this->addFlash('error',"Accès interdit");
+            return $this->redirectToRoute('app_main');
         }
-
-
     }
-
 
 }
